refactor(library/Attempt): moves `DiscriminableOutcome` to new file
